Make Role-specific Category feature unaccessable for unauthorized users

+++ SIMS_CW/Controllers/ManagerCategoryController.cs
@@ -13,16 +13,39 @@ namespace SIMS_CW.Controllers
     public class ManagerCategoryController : Controller
     {
         private DbModel db = new DbModel();
+        private user loggedIn;
 
         // GET: ManagerCategory
         public ActionResult Index()
         {
+            loggedIn = (user)Session["loggedIn"];
+
+            //check logged in?
+            if (loggedIn == null)
+            {
+                return Redirect("~/Home/LoginPage");
+            }
+            if (loggedIn.role.role_id < 1 && loggedIn.role.role_id > 2)
+            {
+                return Redirect("~/Home/DeniedAccess");
+            }
+
             return View(db.categories.ToList());
         }
 
         // GET: ManagerCategory/Details/5
         public ActionResult Details(int? id)
         {
+            //check logged in?
+            if (loggedIn == null)
+            {
+                return Redirect("~/Home/LoginPage");
+            }
+            if (loggedIn.role.role_id < 1 && loggedIn.role.role_id > 2)
+            {
+                return Redirect("~/Home/DeniedAccess");
+            }
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -38,6 +61,16 @@ namespace SIMS_CW.Controllers
         // GET: ManagerCategory/Create
         public ActionResult Create()
         {
+            //check logged in?
+            if (loggedIn == null)
+            {
+                return Redirect("~/Home/LoginPage");
+            }
+            if (loggedIn.role.role_id < 1 && loggedIn.role.role_id > 2)
+            {
+                return Redirect("~/Home/DeniedAccess");
+            }
+
             return View();
         }
 
@@ -48,6 +81,16 @@ namespace SIMS_CW.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "category_id,category_name")] category category)
         {
+            //check logged in?
+            if (loggedIn == null)
+            {
+                return Redirect("~/Home/LoginPage");
+            }
+            if (loggedIn.role.role_id < 1 && loggedIn.role.role_id > 2)
+            {
+                return Redirect("~/Home/DeniedAccess");
+            }
+
             if (ModelState.IsValid)
             {
                 category.created_at = DateTime.Now;
@@ -62,6 +105,16 @@ namespace SIMS_CW.Controllers
         // GET: ManagerCategory/Edit/5
         public ActionResult Edit(int? id)
         {
+            //check logged in?
+            if (loggedIn == null)
+            {
+                return Redirect("~/Home/LoginPage");
+            }
+            if (loggedIn.role.role_id < 1 && loggedIn.role.role_id > 2)
+            {
+                return Redirect("~/Home/DeniedAccess");
+            }
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -81,6 +134,16 @@ namespace SIMS_CW.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "category_id,category_name,created_at")] category category)
         {
+            //check logged in?
+            if (loggedIn == null)
+            {
+                return Redirect("~/Home/LoginPage");
+            }
+            if (loggedIn.role.role_id < 1 && loggedIn.role.role_id > 2)
+            {
+                return Redirect("~/Home/DeniedAccess");
+            }
+
             if (ModelState.IsValid)
             {
                 db.Entry(category).State = EntityState.Modified;
@@ -93,6 +156,16 @@ namespace SIMS_CW.Controllers
         // GET: ManagerCategory/Delete/5
         public ActionResult Delete(int? id)
         {
+            //check logged in?
+            if (loggedIn == null)
+            {
+                return Redirect("~/Home/LoginPage");
+            }
+            if (loggedIn.role.role_id < 1 && loggedIn.role.role_id > 2)
+            {
+                return Redirect("~/Home/DeniedAccess");
+            }
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -110,6 +183,16 @@ namespace SIMS_CW.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            //check logged in?
+            if (loggedIn == null)
+            {
+                return Redirect("~/Home/LoginPage");
+            }
+            if (loggedIn.role.role_id < 1 && loggedIn.role.role_id > 2)
+            {
+                return Redirect("~/Home/DeniedAccess");
+            }
+
             category category = db.categories.Find(id);
             db.categories.Remove(category);
             db.SaveChanges();
